requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
psycopg2-binary>=2.9
gunicorn>=22.0
openpyxl>=3.1
